Create stubs for non-windows code
