+++ pkg/agent/debugspy/placeholder.go
@@ -1,3 +1,1 @@
-// +build !debugspy
-
 package debugspy
