+++ pkg/agent/phpspy/placeholder.go
@@ -1,3 +1,1 @@
-// +build !phpspy
-
 package phpspy
