+++ pkg/agent/pyspy/placeholder.go
@@ -1,5 +1,3 @@
-// +build !pyspy
-
 package pyspy
 
 // TODO: we should probably find a better way of setting this
