+++ pkg/agent/pyspy/pyspy.go
@@ -3,6 +3,7 @@
 // Package pyspy is a wrapper around this library called pyspy written in Rust
 package pyspy
 
+// #cgo windows LDFLAGS: -L../../../third_party/rustdeps/target/release -lrustdeps
 // #cgo darwin LDFLAGS: -L../../../third_party/rustdeps/target/release -lrustdeps
 // #cgo linux LDFLAGS: -L../../../third_party/rustdeps/target/release -lrustdeps
 // #include "../../../third_party/rustdeps/pyspy.h"
