+++ pkg/agent/rbspy/rbspy.go
@@ -3,6 +3,7 @@
 // Package rbspy is a wrapper around this library called rbspy written in Rust
 package rbspy
 
+// #cgo windows LDFLAGS: -L../../../third_party/rustdeps/target/release -lrustdeps
 // #cgo darwin LDFLAGS: -L../../../third_party/rustdeps/target/release -lrustdeps
 // #cgo linux LDFLAGS: -L../../../third_party/rustdeps/target/release -lrustdeps
 // #include "../../../third_party/rustdeps/rbspy.h"
