+++ pkg/util/disk/usage_unix_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package disk
 
 import (
